add languages to info.plist, move extension to correct folder
