+++ lib/base/extensions/locale+fullname.dart
@@ -1,0 +1,13 @@
+import 'dart:ui';
+
+extension FullName on Locale {
+  String fullName() {
+    switch (languageCode) {
+      case 'en':
+        return 'English';
+      case 'de':
+        return 'Deutsch';
+    }
+    return 'fullName not defined';
+  }
+}
