+++ lib/main.dart
@@ -81,15 +81,3 @@ class _AuthenticationRouterState extends ConsumerState<AuthenticationRouter> {
         });
   }
 }
-
-extension FullName on Locale {
-  String fullName() {
-    switch (languageCode) {
-      case 'en':
-        return 'English';
-      case 'de':
-        return 'Deutsch';
-    }
-    return 'fullName not defined';
-  }
-}
